TST non-regression test for auto-memmap / gc bug
